workspace/projects/tg-casino/requirements.txt
python-telegram-bot==21.0
tronpy==0.4.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
cryptography==42.0.0
aiosqlite==0.19.0
redis==5.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
# httpx version managed by python-telegram-bot
